refactor: move skills installer helpers to lib
Share skills CLI installation helpers from lib now that the standalone skill command directory is gone.
